Refactor ExperimentExtension using Strategy pattern
Major refactoring to reduce ExperimentExtension from 1,575 lines to 174 lines
by extracting mode-specific logic into dedicated strategy classes.

Architecture:
- ExperimentConfig: Base interface for mode configurations
- ExperimentModeStrategy: Strategy interface for mode-specific behavior
- MeasureStrategy/ExploreStrategy/OptimizeStrategy: Mode implementations

New packages:
- experiment.measure: MeasureConfig, MeasureStrategy, MeasureInvocationContext, etc.
- experiment.explore: ExploreConfig, ExploreStrategy, ExploreInvocationContext, etc.
- experiment.engine.shared: FactorResolver, ResultRecorder, parameter resolvers

Benefits:
- Mode-agnostic coordinator (no switch/if on mode in extension)
- Symmetric package structure for all three modes
- Testable strategy classes
- Clear separation of concerns
