+++ src/main/java/org/javai/punit/experiment/measure/MeasureInvocationContext.java
@@ -1,0 +1,31 @@
+package org.javai.punit.experiment.measure;
+
+import java.util.List;
+import org.javai.punit.api.ResultCaptor;
+import org.javai.punit.experiment.engine.shared.CaptorParameterResolver;
+import org.junit.jupiter.api.extension.Extension;
+import org.junit.jupiter.api.extension.TestTemplateInvocationContext;
+
+/**
+ * Invocation context for @MeasureExperiment without factors.
+ *
+ * <p>Represents a single sample execution in MEASURE mode when no @FactorSource
+ * is present. Each invocation receives a fresh {@link ResultCaptor}.
+ */
+public record MeasureInvocationContext(
+        int sampleNumber,
+        int totalSamples,
+        String useCaseId,
+        ResultCaptor captor
+) implements TestTemplateInvocationContext {
+
+    @Override
+    public String getDisplayName(int invocationIndex) {
+        return String.format("[%s] sample %d/%d", useCaseId, sampleNumber, totalSamples);
+    }
+
+    @Override
+    public List<Extension> getAdditionalExtensions() {
+        return List.of(new CaptorParameterResolver(captor, null, sampleNumber));
+    }
+}
